util/corsola-relevant-paths.txt firmware-corsola-15194.B-main

TEST=`make -j buildall`

Cq-Depend: chromium:4547220
Force-Relevant-Builds: all

+++ subsys/bluetooth/mesh/pb_adv.c
@@ -14,7 +14,6 @@
 #include "adv.h"
 #include "crypto.h"
 #include "beacon.h"
-#include "host/ecc.h"
 #include "prov.h"
 
 #include "common/bt_str.h"
@@ -136,7 +135,7 @@ static void close_link(enum prov_bearer_link_status status);
 
 static void buf_sent(int err, void *user_data)
 {
-	enum prov_bearer_link_status reason = (enum prov_bearer_link_status)user_data;
+	enum prov_bearer_link_status reason = (enum prov_bearer_link_status)(int)user_data;
 
 	if (atomic_test_and_clear_bit(link.flags, ADV_LINK_CLOSING)) {
 		close_link(reason);
